requirements.txt
requests
fastapi
uvicorn
langchain
openai
faiss-cpu
python-dotenv
pypdf
langchain-openai
langchain-community